pubsub_inference/requirements.txt
tensorflow 
keras
json_numpy
google-cloud-pubsub
google-cloud-storage
google-cloud-bigquery
pandas
numpy